+++ packages/anthropic/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-gui/anthropic
 
+## 0.9.0
+
+### Patch Changes
+
+- @ai-gui/core@0.9.0
+
 ## 0.8.0
 
 ### Patch Changes